ui image assets path extraction
