+++ lib/ui/res/UiImagePaths.dart
@@ -1,3 +1,7 @@
 class UiImagePaths {
-  static const cafe = '';
+  static const String info = 'res/images/info.png';
+  static const String onboardingPage1 = 'res/images/onboarding_page1.png';
+  static const String onboardingPage2 = 'res/images/onboarding_page2.png';
+  static const String onboardingPage3 = 'res/images/onboarding_page3.png';
+  static const String points = 'res/images/onboarding_page3.png';
 }
